Refactoring media screen: activity to fragment
