+++ app/src/main/java/ui/media/MediaFragment.kt
@@ -1,0 +1,50 @@
+package ui.media
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.FragmentMediaBinding
+import com.google.android.material.tabs.TabLayoutMediator
+
+class MediaFragment : Fragment() {
+
+    private lateinit var binding: FragmentMediaBinding
+    private lateinit var tabMediator: TabLayoutMediator
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        binding = FragmentMediaBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        configureViewPager()
+        tabMediator = configureTabMediator()
+        tabMediator.attach()
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        tabMediator.detach()
+    }
+
+    private fun configureViewPager() {
+        binding.viewPager.adapter = MediaViewPagerAdapter(fragment = this)
+    }
+
+    private fun configureTabMediator(): TabLayoutMediator =
+        TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
+            val stringResources = if (position == 0) {
+                R.string.favorite_tracks_title
+            } else R.string.playlists_title
+            tab.text = getString(stringResources)
+        }
+}
