+++ app/src/main/java/ui/media/MediaViewPagerAdapter.kt
@@ -1,18 +1,17 @@
 package ui.media
 
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import ui.media.favoriteTracks.FavoriteTracksFragment
 import ui.media.playlists.PlaylistsFragment
 
 class MediaViewPagerAdapter(
-    fragmentManager: FragmentManager,
-    lifecycle: Lifecycle
-) : FragmentStateAdapter(fragmentManager, lifecycle) {
+    fragment: Fragment
+) : FragmentStateAdapter(fragment) {
+
     override fun getItemCount() = 2
 
     override fun createFragment(position: Int): Fragment =
-        if (position == 0) FavoriteTracksFragment.newInstance() else PlaylistsFragment.newInstance()
+        if (position == 0) FavoriteTracksFragment.newInstance()
+        else PlaylistsFragment.newInstance()
 }
