--- app/src/main/res/layout/activity_media.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <ImageView
-            android:id="@+id/backButton"
-            style="@style/BackButtonNavigationBarStyle" />
-
-        <TextView
-            style="@style/SecondaryActivityTitleStyle"
-            android:layout_height="match_parent"
-            android:layout_marginStart="12dp"
-            android:text="@string/media_button" />
-    </LinearLayout>
-
-    <com.google.android.material.tabs.TabLayout
-        android:id="@+id/mediaTabLayout"
-        style="@style/TabLayoutStyle" />
-
-    <androidx.viewpager2.widget.ViewPager2
-        android:id="@+id/mediaViewPager"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-</LinearLayout>
